extract more private scanner funcs
